Rename class & exposed method, create store.js
