+++ src/index.js
@@ -1,7 +1,7 @@
 import EventEmitter from 'events';
 
 import { setPromisesDependency, getPromisesDependency } from './utils';
-import { main, readyFreddy, getEnv, getValue, setValue, pushRecord, getOrCreateStore, apifyClient, call, getMemoryInfo } from './actor';
+import { main, readyFreddy, getEnv, getValue, setValue, pushRecord, openKeyValueStore, apifyClient, call, getMemoryInfo } from './actor';
 import { launchPuppeteer } from './puppeteer';
 import { browse, launchWebDriver } from './webdriver';
 
@@ -14,7 +14,7 @@ const Apify = {
     getValue,
     setValue,
     pushRecord,
-    getOrCreateStore,
+    openKeyValueStore,
     call,
     readyFreddy,
     setPromisesDependency,
