+++ test/actor.js
@@ -1153,35 +1153,6 @@ describe('Apify.getOrCreateStore()', () => {
             expect(() => { store.setValue('key', 'value', { contentType: '' }); }).to.throw(Error, contTypeStringErrMsg);
         });
     });
-
-    const testStore = storeFabric('my-test-store');
-    it('returns null from a non-existent key in getValue method', () => {
-        testStore.then((store) => {
-            return store.getValue('INPUT').then((input) => {
-                return expect(input).to.eql(null);
-            });
-        });
-    });
-
-    const setValueStore = storeFabric('my-setValue-store');
-    it('returns STATE value after calling setValue method', () => {
-        const mockState = JSON.stringify({
-            testString: 'String',
-            testNumber: 3e6,
-            testArray: [1, 2, 3, 4],
-            testObject: {
-                key: 'value',
-            },
-        });
-        setValueStore.then((store) => {
-            return store.setValue('STATE', mockState).then(() => {
-                return store.getValue('STATE').then((state) => {
-                    expect(state).to.be.a('string');
-                    expect(state).to.eql(mockState);
-                }).catch(error => console.log('Oh no 2!', error));
-            });
-        });
-    });
 });
 
 describe('Apify.events', () => {
